fix: apply CodeRabbit auto-fixes
Add onboarding_setup_journey to redirect input type and clear wallet session
flags from the shared auth logout and account-switch paths.
