+++ packages/authentication/src/index.ts
@@ -1,4 +1,5 @@
 export * from './auth-return-path';
+export * from './onboarding-wallet-session';
 export * from './resolve-post-auth-redirect-path';
 export * from './shared/types';
 export * from './provider';
