+++ packages/epics/src/common/onboarding-wallet-handoff.ts
@@ -1,8 +1,13 @@
 'use client';
 
+export {
+  clearOnboardingWalletSessionActive,
+  isOnboardingWalletSessionActive,
+  markOnboardingWalletSessionActive,
+} from '@hypha-platform/authentication';
+
 const COMPLETED_SLUGS_KEY = 'hypha:onboarding-wallet-handoff:slugs:v1';
 const HANDLED_PAYLOADS_KEY = 'hypha:onboarding-wallet-handoff:payloads:v1';
-const WALLET_SESSION_ACTIVE_KEY = 'hypha:onboarding-wallet-session:active:v1';
 
 function readStringSet(key: string): Set<string> {
   if (typeof window === 'undefined') return new Set();
@@ -25,33 +30,6 @@ function writeStringSet(key: string, values: Set<string>): void {
     window.sessionStorage.setItem(key, JSON.stringify([...values]));
   } catch {
     // ignore quota / private mode
-  }
-}
-
-export function isOnboardingWalletSessionActive(): boolean {
-  if (typeof window === 'undefined') return false;
-  try {
-    return window.sessionStorage.getItem(WALLET_SESSION_ACTIVE_KEY) === '1';
-  } catch {
-    return false;
-  }
-}
-
-export function markOnboardingWalletSessionActive(): void {
-  if (typeof window === 'undefined') return;
-  try {
-    window.sessionStorage.setItem(WALLET_SESSION_ACTIVE_KEY, '1');
-  } catch {
-    // ignore quota / private mode
-  }
-}
-
-export function clearOnboardingWalletSessionActive(): void {
-  if (typeof window === 'undefined') return;
-  try {
-    window.sessionStorage.removeItem(WALLET_SESSION_ACTIVE_KEY);
-  } catch {
-    // ignore
   }
 }
